+++ volmdlr/edges.py
@@ -289,6 +289,45 @@ class Edge(dc.DessiaObject):
                 break
         return new_split_edge
 
+    def point_distance_to_edge(self, point):
+        """
+        Calculates the distance from a given point to an edge.
+
+        :param point: point.
+        :return: distance to edge.
+        """
+        best_distance = math.inf
+        abscissa1 = 0
+        abscissa2 = self.abscissa(self.end)
+        distance = best_distance
+        point1_ = None
+        point2_ = None
+        linesegment_class_ = getattr(sys.modules[__name__], 'LineSegment' + self.__class__.__name__[-2:])
+        while True:
+            discretized_points_between_1_2 = []
+            for abscissa in npy.linspace(abscissa1, abscissa2, num=8):
+                abscissa_point = self.point_at_abscissa(abscissa)
+                if not volmdlr.core.point_in_list(abscissa_point, discretized_points_between_1_2):
+                    discretized_points_between_1_2.append(abscissa_point)
+            if not discretized_points_between_1_2:
+                break
+            distance = point.point_distance(discretized_points_between_1_2[0])
+            for point1, point2 in zip(discretized_points_between_1_2[:-1], discretized_points_between_1_2[1:]):
+                line = linesegment_class_(point1, point2)
+                dist = line.point_distance(point)
+                if dist < distance:
+                    point1_ = point1
+                    point2_ = point2
+                    distance = dist
+            if not point1_ or math.isclose(distance, best_distance, abs_tol=1e-6):
+                break
+            abscissa1 = self.abscissa(point1_)
+            abscissa2 = self.abscissa(point2_)
+            best_distance = distance
+            if math.isclose(abscissa1, abscissa2, abs_tol=1e-6):
+                break
+        return distance
+
 
 class Line(dc.DessiaObject):
     """
@@ -1327,12 +1366,12 @@ class BSplineCurve(Edge):
                 abs1 = self.abscissa(linesegment.start)
                 abs2 = self.abscissa(linesegment.end)
                 list_abscissas = list(new_abscissa for new_abscissa in npy.linspace(abs1, abs2, 1000))
-                intersection = self.select_intersection_point(list_abscissas, intersections)
+                intersection = self.select_intersection_point(list_abscissas, intersections, line)
                 list_intersections.append(intersection)
             initial_abscissa += linesegment.length()
         return list_intersections
 
-    def select_intersection_point(self, list_abscissas, intersections):
+    def select_intersection_point(self, list_abscissas, intersections, line, abs_tol: float = 1e-7):
         """
         Select closest point in curve to intersection point obtained with discretized linesegment.
 
@@ -1344,6 +1383,8 @@ class BSplineCurve(Edge):
         intersection = None
         for i_abscissa in list_abscissas:
             point_in_curve = BSplineCurve.point_at_abscissa(self, i_abscissa)
+            if line.point_distance(point_in_curve) <= abs_tol:
+                return point_in_curve
             dist = point_in_curve.point_distance(intersections[0])
             if dist < distance:
                 distance = dist
@@ -1552,7 +1593,7 @@ class Line2D(Line):
             a new figure is created.
         :type ax: matplotlib.axes._subplots.AxesSubplot, optional
         :param edge_style: data class instance, containing all parameters needed to plot Line 2D.
-        :return: The matplotlib axis.
+        :return: The Matplotlib axis.
         :rtype: matplotlib.axes._subplots.AxesSubplot
         """
         if ax is None:
@@ -1991,36 +2032,7 @@ class BSplineCurve2D(BSplineCurve):
         :param point: point 2d.
         :return: distance.
         """
-        best_distance = math.inf
-        abscissa1 = 0
-        abscissa2 = self.abscissa(self.end)
-        distance = best_distance
-        point1_ = None
-        point2_ = None
-        while True:
-            discretized_points_between_1_2 = []
-            for abscissa in npy.linspace(abscissa1, abscissa2, num=8):
-                abscissa_point = self.point_at_abscissa(abscissa)
-                if not volmdlr.core.point_in_list(abscissa_point, discretized_points_between_1_2):
-                    discretized_points_between_1_2.append(abscissa_point)
-            if not discretized_points_between_1_2:
-                break
-            distance = point.point_distance(discretized_points_between_1_2[0])
-            for point1, point2 in zip(discretized_points_between_1_2[:-1], discretized_points_between_1_2[1:]):
-                line = LineSegment2D(point1, point2)
-                dist = line.point_distance(point)
-                if dist < distance:
-                    point1_ = point1
-                    point2_ = point2
-                    distance = dist
-            if not point1_ or math.isclose(distance, best_distance, abs_tol=1e-6):
-                break
-            abscissa1 = self.abscissa(point1_)
-            abscissa2 = self.abscissa(point2_)
-            best_distance = distance
-            if math.isclose(abscissa1, abscissa2, abs_tol=1e-6):
-                break
-        return distance
+        return self.point_distance_to_edge(point)
 
     def nearest_point_to(self, point):
         """
@@ -3176,6 +3188,18 @@ class Arc2D(Arc):
         return ax
 
     def to_3d(self, plane_origin, x, y):
+        """
+        Transforms the arc 2D into a 3D arc.
+
+        :param plane_origin: The origin of plane to draw the arc 3D.
+        :type plane_origin: volmdlr.Point3D
+        :param x: First direction of the plane
+        :type x: volmdlr.Vector3D
+        :param y: Second direction of the plane.
+        :type y: volmdlr.Vector3D
+        :return: A 3D arc.
+        :type: Arc3D.
+        """
         point_start = self.start.to_3d(plane_origin, x, y)
         point_interior = self.interior.to_3d(plane_origin, x, y)
         point_end = self.end.to_3d(plane_origin, x, y)
@@ -3445,6 +3469,18 @@ class FullArc2D(Arc2D):
         return False
 
     def to_3d(self, plane_origin, x, y):
+        """
+        Transforms the full arc 2D into a 3D full arc.
+
+        :param plane_origin: The origin of plane to draw the full arc 3D.
+        :type plane_origin: volmdlr.Point3D
+        :param x: First direction of the plane
+        :type x: volmdlr.Vector3D
+        :param y: Second direction of the plane.
+        :type y: volmdlr.Vector3D
+        :return: A 3D full arc.
+        :type: Full Arc 3D.
+        """
         center = self.center.to_3d(plane_origin, x, y)
         start = self.start.to_3d(plane_origin, x, y)
         z = x.cross(y)
@@ -3639,46 +3675,36 @@ class ArcEllipse2D(Edge):
         interior_new = frame.global_to_local_coordinates(self.interior)
         center_new = frame.global_to_local_coordinates(self.center)
         self._bounding_rectangle = None
+        self._reverse = None
 
-        def theta_a_b(start_, iterior_, end_, center_):
-            """
-            Calculates the major, minor and theta.
-
-            From : https://math.stackexchange.com/questions/339126/how-to-draw-an-ellipse-if-a- \
-            center-and-3-arbitrary-points-on-it-are-given.
-            theta= ellipse's inclination angle related to the horizontal
-            (clockwise), A=semi major axis, B=semi minor axis.
-
-            """
-            x_start, y_start, x_interior, y_interior, x_end, y_end = start_[0] - center_[0], start_[1] - center_[1], \
-                iterior_[0] - center_[0], iterior_[1] - center_[
-                                                                         1], end_[0] - center_[0], end_[1] - center_[1]
-            matrix_a = npy.array(([x_start ** 2, y_start ** 2, 2 * x_start * y_start],
-                                  [x_interior ** 2, y_interior ** 2, 2 * x_interior * y_interior],
-                                  [x_end ** 2, y_end ** 2, 2 * x_end * y_end]))
-            inv_matrix_a = npy.linalg.inv(matrix_a)
-            matriz_one = npy.array(([1],
-                                    [1],
-                                    [1]))
-            vector_c = npy.dot(inv_matrix_a, matriz_one)
-            theta = 0.5 * math.atan(2 * vector_c[2] / (vector_c[1] - vector_c[0]))
-            c1 = vector_c[0] + vector_c[1]
-            c2 = (vector_c[1] - vector_c[0]) / math.cos(2 * theta)
-            gdaxe = math.sqrt((2 / (c1 - c2)))
-            ptax = math.sqrt((2 / (c1 + c2)))
-            return theta, gdaxe, ptax
-
-        if start.is_close(end):
-            extra_new = frame.global_to_local_coordinates(self.extra)
-            theta, major_axis, minor_axis = theta_a_b(start_new, extra_new, interior_new,
-                                                      center_new)
+        # vector_center_interior = interior_new - center_new
+        vector_center_start = start_new - center_new
+        vector_center_end = end_new - center_new
+        if vector_center_start.norm() >= vector_center_end.norm():
+            x1 = start_new.x - center_new.x
+            y1 = start_new.y - center_new.y
+            x2 = end_new.x - center_new.x
+            y2 = end_new.y - center_new.y
         else:
-            theta, major_axis, minor_axis = theta_a_b(start_new, interior_new, end_new,
-                                                      center_new)
-
+            x2 = start_new.x - center_new.x
+            y2 = start_new.y - center_new.y
+            x1 = end_new.x - center_new.x
+            y1 = end_new.y - center_new.y
+        if vector_center_start.is_colinear_to(vector_center_end) or abs(x1) == abs(x2):
+            x2 = interior_new.x - center_new.x
+            y2 = interior_new.y - center_new.y
+            if abs(x1) == abs(x2):
+                raise ValueError(f"Interior point{interior} is not valid. Try specifying another interior point.")
+        minor_axis = math.sqrt((x1**2 * y2**2 - x2**2 * y1**2) / (x1**2 - x2**2))
+        if (1 - (y1 ** 2 / minor_axis ** 2)) == 0.0:
+            if abs(y1) != abs(y2) and abs(interior_new.y) != abs(y2):
+                x1 = interior_new.x - center_new.x
+                y1 = interior_new.y - center_new.y
+            else:
+                raise NotImplementedError
+        major_axis = math.sqrt(x1 ** 2 / (1 - (y1 ** 2 / minor_axis ** 2)))
         self.major_axis = major_axis
         self.minor_axis = minor_axis
-        self.theta = theta
 
         # Angle pour start
         u1, u2 = start_new.x / self.major_axis, start_new.y / self.minor_axis
@@ -3735,8 +3761,9 @@ class ArcEllipse2D(Edge):
 
         :return: arc ellipse 2d's length
         """
-        length = self.abscissa(self.end)
-        return length
+        if not self._length:
+            self._length = self.abscissa(self.end)
+        return self._length
 
     def point_belongs(self, point, abs_tol: float = 1e-6):
         """
@@ -3746,19 +3773,81 @@ class ArcEllipse2D(Edge):
         :param abs_tol: tolerance applied during calculations
         :return: True if the point belongs, False otherwise
         """
+        if self.start.is_close(point, abs_tol) or self.end.is_close(point, abs_tol):
+            return True
         if not math.isclose((point.x - self.center.x) ** 2 / self.major_axis ** 2 +
                             (point.y - self.center.y) ** 2 / self.minor_axis ** 2, 1, abs_tol=abs_tol) and not \
                 math.isclose((point.x - self.center.x) ** 2 / self.minor_axis ** 2 +
                              (point.y - self.center.y) ** 2 / self.major_axis ** 2, 1, abs_tol=abs_tol):
             return False
-        new_point = self.frame.global_to_local_coordinates(point)
-        u1, u2 = new_point.x / self.major_axis, new_point.y / self.minor_axis
-        angle_new_point = volmdlr.geometry.sin_cos_angle(u1, u2)
-        if self.angle_start < self.angle_end and self.angle_end >= angle_new_point >= self.angle_start:
-            return True
-        if self.angle_start > self.angle_end and self.angle_end <= angle_new_point <= self.angle_start:
+        clockwise_arcellipse = self.reverse() if self.is_trigo else self
+        vector_start = clockwise_arcellipse.start - clockwise_arcellipse.center
+        vector_end = clockwise_arcellipse.end - clockwise_arcellipse.center
+        vector_point = point - clockwise_arcellipse.center
+        arc_angle = volmdlr.geometry.clockwise_angle(vector_start, vector_end)
+        point_start_angle = volmdlr.geometry.clockwise_angle(vector_start, vector_point)
+        point_end_angle = volmdlr.geometry.clockwise_angle(vector_point, vector_end)
+        if math.isclose(arc_angle, point_start_angle + point_end_angle, abs_tol=1e-5):
             return True
         return False
+
+    def valid_abscissa_start_end_angle(self, angle_abscissa):
+        angle_start = self.angle_start
+        angle_end = angle_abscissa
+        if self.angle_start > angle_abscissa >= self.angle_end:
+            if angle_abscissa >= 0.0:
+                angle_abscissa += 2 * math.pi
+                angle_end = angle_abscissa
+            else:
+                angle_start = angle_abscissa
+                angle_end = self.angle_start
+        elif self.angle_start > self.angle_end >= angle_abscissa:
+            angle_start = self.angle_start - 2 * math.pi
+        return angle_start, angle_end
+
+    def point_at_abscissa(self, abscissa):
+        if math.isclose(abscissa, 0.0, abs_tol=1e-6):
+            return self.start
+        if math.isclose(abscissa, self.length(), abs_tol=1e-6):
+            return self.end
+        if not self.is_trigo:
+            arc_ellipse_trigo = self.reverse()
+            new_abscissa = self.length() - abscissa
+            return arc_ellipse_trigo.point_at_abscissa(new_abscissa)
+        discretized_points = self.discretization_points(number_points=100)
+        aproximation_abscissa = 0
+        aproximation_point = None
+        for point1, point2 in zip(discretized_points[:-1], discretized_points[1:]):
+            dist1 = point1.point_distance(point2)
+            if aproximation_abscissa + dist1 > abscissa:
+                aproximation_point = point1
+                break
+            aproximation_abscissa += dist1
+        initial_point = self.frame.global_to_local_coordinates(aproximation_point)
+        u1, u2 = initial_point.x / self.major_axis, initial_point.y / self.minor_axis
+        initial_angle = volmdlr.geometry.sin_cos_angle(u1, u2)
+        angle_start, initial_angle = self.valid_abscissa_start_end_angle(initial_angle)
+        def ellipse_arc_length(theta):
+            return math.sqrt((self.major_axis ** 2) * math.sin(theta) ** 2 +
+                             (self.minor_axis ** 2) * math.cos(theta) ** 2)
+        abscissa_angle = None
+        iter_counter = 0
+        increment_factor = 1e-5
+        while True:
+            res, _ = scipy_integrate.quad(ellipse_arc_length, angle_start, initial_angle)
+            if math.isclose(res, abscissa, abs_tol=1e-5):
+                abscissa_angle = initial_angle
+                break
+            if res > abscissa:
+                if iter_counter == 0:
+                    increment_factor = -1e-5
+                else:
+                    raise NotImplementedError
+            initial_angle += increment_factor
+            iter_counter += 1
+        x = self.major_axis * math.cos(abscissa_angle)
+        y = self.minor_axis * math.sin(abscissa_angle)
+        return self.frame.local_to_global_coordinates(volmdlr.Point2D(x, y))
 
     def abscissa(self, point: volmdlr.Point2D):
         """
@@ -3767,19 +3856,30 @@ class ArcEllipse2D(Edge):
         :param point: point for calculating abscissa
         :return: a float, between 0 and the arc ellipse 2d's length
         """
-        if self.point_belongs(point):
-            angle_abscissa = volmdlr.geometry.clockwise_angle(point - self.center, self.major_dir)
-            angle_start = self.angle_start
-            angle_end = angle_abscissa
-            if self.angle_start > angle_abscissa > self.angle_end:
-                angle_start = angle_abscissa
-                angle_end = self.angle_start
+        if self.start.is_close(point):
+            return 0.0
+        if self.end.is_close(point):
+            if self._length:
+                return self._length
+            if not self.is_trigo:
+                arc_ellipse_trigo = self.reverse()
+                abscissa_end = arc_ellipse_trigo.abscissa(self.start)
+                return abscissa_end
 
-            def arc_length(theta):
+        if self.point_belongs(point):
+            if not self.is_trigo:
+                arc_ellipse_trigo = self.reverse()
+                abscissa_point = arc_ellipse_trigo.abscissa(point)
+                return self.length() - abscissa_point
+            new_point = self.frame.global_to_local_coordinates(point)
+            u1, u2 = new_point.x / self.major_axis, new_point.y / self.minor_axis
+            angle_abscissa = volmdlr.geometry.sin_cos_angle(u1, u2)
+            angle_start, angle_end = self.valid_abscissa_start_end_angle(angle_abscissa)
+            def ellipse_arc_length(theta):
                 return math.sqrt((self.major_axis ** 2) * math.sin(theta) ** 2 +
                                  (self.minor_axis ** 2) * math.cos(theta) ** 2)
 
-            res, _ = scipy_integrate.quad(arc_length, angle_start, angle_end)
+            res, _ = scipy_integrate.quad(ellipse_arc_length, angle_start, angle_end)
             return res
         raise ValueError(f'point {point} does not belong to ellipse')
 
@@ -3845,6 +3945,10 @@ class ArcEllipse2D(Edge):
         elif self.angle_start == self.angle_end:
             angle_start = 0
             angle_end = 2 * math.pi
+        elif self.angle_end > self.angle_start and not self.is_trigo:
+            angle_start = self.angle_end - 2 * math.pi
+            angle_end = self.angle_start
+            is_trigo = False
         else:
             angle_end = self.angle_end
             angle_start = self.angle_start
@@ -3873,7 +3977,7 @@ class ArcEllipse2D(Edge):
         :param y: Second direction of the plane.
         :type y: volmdlr.Vector3D
         :return: A 3D arc of ellipse.
-        :rtype: ArcEllipse3D
+        :type: ArcEllipse3D.
         """
         point_start3d = self.start.to_3d(plane_origin, x, y)
         point_interior3d = self.interior.to_3d(plane_origin, x, y)
@@ -3891,6 +3995,13 @@ class ArcEllipse2D(Edge):
                             point_center3d, new_major_dir, extra3d, name=self.name)
 
     def plot(self, ax=None, edge_style: EdgeStyle = EdgeStyle()):
+        """
+        Plot arc-ellipse 2d using matplotlob.
+
+        :param ax: Matplotlib plot if there exists any.
+        :param edge_style: edge styles.
+        :return: Matplotlib plot
+        """
         if ax is None:
             _, ax = plt.subplots()
 
@@ -3915,8 +4026,10 @@ class ArcEllipse2D(Edge):
         raise NotImplementedError
 
     def reverse(self):
-        return self.__class__(self.end.copy(), self.interior.copy(), self.start.copy(),
+        if not self._reverse:
+            self._reverse = self.__class__(self.end.copy(), self.interior.copy(), self.start.copy(),
                               self.center.copy(), self.major_dir.copy(), self.name)
+        return self._reverse
 
     def line_intersections(self, line2d: Line2D):
         """
@@ -3971,6 +4084,28 @@ class ArcEllipse2D(Edge):
                                 major_dir)
         raise ValueError('Side should be \'new\' \'old\'')
 
+    def translation(self, offset: volmdlr.Vector2D):
+        """
+        Translates the Arc ellipse given an offset vector.
+
+        :param offset: offset vector
+        :return: new translated arc ellipse 2d.
+        """
+        new_start = self.start.translation(offset)
+        new_end = self.end.translation(offset)
+        new_interior = self.interior.translation(offset)
+        new_center = self.center.translation(offset)
+        return ArcEllipse2D(new_start, new_interior, new_end, new_center, self.major_dir)
+
+    def point_distance(self, point):
+        """
+        Calculates the distance from a given point to an Arc Ellipse 2d.
+
+        :param point: point 2d.
+        :return: distance.
+        """
+        return self.point_distance_to_edge(point)
+
     def straight_line_point_belongs(self, point):
         """
         Verifies if a point belongs to the surface created by closing the edge.
@@ -3982,6 +4117,23 @@ class ArcEllipse2D(Edge):
         raise NotImplementedError(f'the straight_line_point_belongs method must be'
                                   f' overloaded by {self.__class__.__name__}')
 
+    def split(self, split_point):
+        """
+        Splits arc-elipse at a given point.
+
+        :param split_point: splitting point.
+        :return: list of two Arc-Ellipse.
+        """
+        if split_point.is_close(self.start, 1e-6):
+            return [None, self.copy()]
+        if split_point.is_close(self.end, 1e-6):
+            return [self.copy(), None]
+        abscissa = self.abscissa(split_point)
+        return [self.__class__(self.start, self.point_at_abscissa(0.5 * abscissa), split_point,
+                               self.center.copy(), self.major_dir.copy()),
+                self.__class__(split_point, self.point_at_abscissa(
+                    (self.abscissa(self.end) - abscissa) * 0.5 + abscissa),
+                               self.end, self.center.copy(), self.major_dir.copy())]
 
 class FullArcEllipse(Edge):
     """
@@ -4004,7 +4156,8 @@ class FullArcEllipse(Edge):
         Calculates the length of the ellipse.
 
         Ramanujan's approximation for the perimeter of the ellipse.
-        P = math.pi * (a + b) [ 1 + (3h) / (10 + √(4 - 3h) ) ], where h = (a - b)**2/(a + b)**2
+        P = math.pi * (a + b) [ 1 + (3h) / (10 + √(4 - 3h) ) ], where h = (a - b)**2/(a + b)**2.
+
         :return: Perimeter of the ellipse
         :rtype: float
         """
@@ -4127,6 +4280,18 @@ class FullArcEllipse2D(FullArcEllipse, ArcEllipse2D):
         return discretization_points
 
     def to_3d(self, plane_origin, x, y):
+        """
+        Transforms the full arc of ellipse 2D into a 3D full arc of ellipse.
+
+        :param plane_origin: The origin of plane to draw the full arc of ellipse 3D.
+        :type plane_origin: volmdlr.Point3D
+        :param x: First direction of the plane
+        :type x: volmdlr.Vector3D
+        :param y: Second direction of the plane.
+        :type y: volmdlr.Vector3D
+        :return: A 3D full arc of ellipse.
+        :rtype: FullArcEllipse3D
+        """
         point_start_end3d = self.start_end.to_3d(plane_origin, x, y)
         point_center3d = self.center.to_3d(plane_origin, x, y)
 
@@ -4526,6 +4691,11 @@ class LineSegment3D(LineSegment):
                 and self.end == other_linesegment3d.end)
 
     def _bounding_box(self):
+        """
+        Calculates the bounding box for a line segment 3D.
+
+        :return: Bounding box for line segment 3d.
+        """
 
         xmin = min(self.start.x, self.end.x)
         xmax = max(self.start.x, self.end.x)
@@ -4561,12 +4731,26 @@ class LineSegment3D(LineSegment):
         return distance
 
     def plane_projection2d(self, center, x, y):
+        """
+        Calculates the projection of a line segment 3d on to a plane.
+
+        :param center: plane center.
+        :param x: plane u direction.
+        :param y: plane v direction.
+        :return: line segment 3d.
+        """
         start, end = self.start.plane_projection2d(center, x, y), self.end.plane_projection2d(center, x, y)
         if not start.is_close(end):
             return LineSegment2D(start, end)
         return None
 
     def line_intersections(self, line):
+        """
+        Gets the intersection between a line segment 3d and line3D.
+
+        :param line: other line.
+        :return: a list with the intersection points.
+        """
         line_self = self.to_line()
         if line_self.skew_to(line):
             return []
@@ -4576,6 +4760,12 @@ class LineSegment3D(LineSegment):
         return []
 
     def linesegment_intersections(self, linesegment):
+        """
+        Gets the intersection between a line segment 3d and another line segment 3D.
+
+        :param linesegment: other line segment.
+        :return: a list with the intersection points.
+        """
         line1 = self.to_line()
         line2 = linesegment.to_line()
         intersection = line1.intersection(line2)